+++ src/actions/cms/articleSections/update.ts
@@ -5,7 +5,7 @@ import prisma from '@/prisma'
 import { createActionError, createPrismaActionError } from '@/actions/error'
 import { createCmsImage } from '@/server/cms/images/create'
 import { createCmsParagraph } from '@/server/cms/paragraphs/create'
-import { createCmsLink } from '@/actions/cms/links/create'
+import { createCmsLink } from '@/server/cms/links/create'
 import type { ImageSize, ArticleSection, Position } from '@prisma/client'
 import type { ExpandedArticleSection } from './Types'
 import type { ActionReturn } from '@/actions/Types'
