+++ src/actions/cms/links/create.ts
@@ -1,18 +1,9 @@
 'use server'
-import { createPrismaActionError } from '@/actions/error'
-import prisma from '@/prisma'
 import type { ActionReturn } from '@/actions/Types'
 import type { CmsLink } from '@prisma/client'
+import { createCmsLink } from '@/server/cms/links/create'
 
-export async function createCmsLink(name: string): Promise<ActionReturn<CmsLink>> {
-    try {
-        const cmsLink = await prisma.cmsLink.create({
-            data: {
-                name
-            }
-        })
-        return { success: true, data: cmsLink }
-    } catch (error) {
-        return createPrismaActionError(error)
-    }
+export async function createCmsLinkAction(name: string): Promise<ActionReturn<CmsLink>> {
+    //TODO: Auth on permission to create cms
+    return await createCmsLink(name)
 }
