+++ src/actions/cms/paragraphs/read.ts
@@ -1,7 +1,6 @@
 'use server'
 import { createCmsParagraph } from '@/server/cms/paragraphs/create'
-import { createActionError, createPrismaActionError } from '@/actions/error'
-import prisma from '@/prisma'
+import { createActionError } from '@/actions/error'
 import { SpecialCmsParagraph } from '@prisma/client'
 import type { ActionReturn } from '@/actions/Types'
 import type { CmsParagraph } from '@prisma/client'
